Major restructuring
The problem was that the project was growing too big; it needed to be
split into a more modular system where everything followed the
single-responsibility principle. This patch accomplishes this by
defining a new project structure and separating the book's typeset
raw notes. The structure of the narrative has not changed drastically,
though certain areas have been rearranged. Rather, the project itself is
organized in a different way to allow for more growth.